Refactor: Consolidate agentic instrumentation into instrumentation.ts
